Created decoders for funct3 and funct7 bit fields
The decoder for funct3 (funct3_decode) is built upon the R and I type decoders. It uses the I type decoder to distinguish between the various load and store instructions (given they use immediates) so far. Given that the logical and arithmetic operations within the R type instructions and I type instructions follow the same encoding in funct3, the R type decoder is used to distinguish between various logical and arithmetic instructions, and  is used for I and R type instructions.

The funct7 decoder is needed to distinguish between logical and arithmetic right shifts, as well as, addition and subtraction.