Created an encoder for the AluCtrl signal
This encoder uses the information on what kind of logical or arithmetic instruction is currently executing, or, if the current instruction is a branch or store. It produces a 4-bit signal encoding this information to control the ALU operation.

The need for including a branch and store input signals comes from the need to perform subtraction or addition respectively, on a base register when those instructions execute.