Created control unit top sheet
The control main sheet represents the full control unit system and uses the ROM to fetch instructions that are processed by the control path. It uses the immSrc output of the control path to form  a 32-bit sign extended immediate using the sign extend module and outputs the other control path outputs. The control main sheet also outputs the source and destination registers by slicing the instruction word as required.