Changed the Sign Extension unit operation
Now the sign extender will intake a 20-bit immediate generated by the ImmDecode decoder, and a extend type signal, which determines if the output is sign extended to 32-bits, or if its just the 20-bit immediate followed by 12 zeroes (needed for upper instructions). Previously in v1, this unit only sign extended the immediate.