added logic to compute MemControl
Mem Control is a signal computed by the control unit that will help determine the value of the immediate offset when calculating a given memory address. This is because the value of the immediate in a store or load instruction, that isn't a load/store word, will require a modification to the immediate offset and to the value written to the destination register(load case) in order to support byte-addressing.