Changed ImmDecode
Changed ImmDecode so that it performs the immediate operand decoding and sign extension in one go rather than having a sign extension unit.

This was done as the B and J type instructions would be 13 bits and 21 bits wide respectively (due to lsb being 0). If the extension unit and immdecode were kept separate, additional control signals would be needed to organize the bits of the sign extended immediate operand when there is a branch or JAL instruction. To avoid additional control logic requirements, the module was made self contained and now requires the instruction type, sub-type and the actual instruction word to compute the extended 32-bit immediate.